# web/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Referral API",
        "description": "This is API Microservice Referral Program",
        "contact": {
            "name": "Support Team",
            "email": "admin@sumra.net"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/tests/tools/data-encrypt": {
            "post": {
                "tags": [
                    "Tools"
                ],
                "summary": "Join new user to referrer",
                "description": "Send encryption data",
                "operationId": "App\\Http\\Controllers\\ToolsController::dataEncrypt",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "version_key": {
                                        "description": "Version Key",
                                        "type": "string",
                                        "example": "66981685"
                                    },
                                    "data": {
                                        "description": "Encrypt data",
                                        "type": "text",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success send data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "default": [
                            "ManagerRead",
                            "User",
                            "ManagerWrite"
                        ]
                    }
                ],
                "x-auth-type": "Application & Application User",
                "x-throttling-tier": "Unlimited",
                "x-wso2-application-security": {
                    "security-types": [
                        "oauth2"
                    ],
                    "optional": "false"
                }
            }
        },
        "/tests/tools/data-decrypt": {
            "post": {
                "tags": [
                    "Tools"
                ],
                "summary": "Join new user to referrer",
                "description": "Send encryption data",
                "operationId": "App\\Http\\Controllers\\ToolsController::dataDecrypt",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "version_key": {
                                        "description": "Version Key",
                                        "type": "string",
                                        "example": "66981685"
                                    },
                                    "data": {
                                        "description": "Encrypt data",
                                        "type": "text",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success send data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "default": [
                            "ManagerRead",
                            "User",
                            "ManagerWrite"
                        ]
                    }
                ],
                "x-auth-type": "Application & Application User",
                "x-throttling-tier": "Unlimited",
                "x-wso2-application-security": {
                    "security-types": [
                        "oauth2"
                    ],
                    "optional": "false"
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "default": {
                "type": "oauth2",
                "description": "Auth Scheme",
                "flows": {
                    "implicit": {
                        "authorizationUrl": "https://is.sumra.net/oauth2",
                        "scopes": {
                            "ManagerRead": "Manager can read",
                            "User": "User access",
                            "ManagerWrite": "Manager can write"
                        }
                    }
                }
            }
        }
    }
}